requirements.txt
langgraph>=0.2.60
langchain>=0.3.20
langchain-community>=0.3.20
langchain-google-genai>=4.2.1
langchain-chroma>=0.2.0
chromadb>=0.5.23
pypdf>=5.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
duckduckgo-search>=7.1.1
tavily-python>=0.5.0
pydantic>=2.10.0
httpx>=0.28.1
langchain-text-splitters>=0.3.6
